cpufreq: imx6q: Fix clock enable balance
For changing the cpu frequency the i.MX6q has to be switched to some
intermediate clock during the PLL reprogramming. The driver tries
to be clever to keep the enable count correct but gets it wrong. If
the cpufreq is increased it calls clk_disable_unprepare twice
on pll2_pfd2_396m. This puts all other devices which get their clock
from pll2_pfd2_396m into a nonworking state.

Fix this by removing the clk enabling/disabling altogether since the
clk core will do this automatically during a reparent.

[shawn.guo: The driver assumes that cpu always boots with 800MHz, and
 the first transition on 400MHz must be entering 400MHz setpoint.  But
 it turns out that's not always the case.  In some system, cpu boots up
 at 400MHz and then the first transition on 400MHz will be leaving
 400MHz setpoint rather than entering.  In this case, the use count of
 pll2_pfd2_396m will be wrong.  Anyway, since clock framework has been
 nicely handling the clk enabling during reparent, we should not need
 to worry about it.]

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Shawn Guo <shawn.guo@freescale.com>