Revert "F/ipu : Remove function that changes the timings for controlling HDMI signals"
This reverts commit a90cd684ef0b1c737ea645f323c91baff7404917.

I have noticed on one of our screens that even with removing "ipu
restrictions" some video modes are not working with the IPU hardware
unit.

Having the restrictions still makes other tv not proberly work with some
video modes.

There are 2 possibilities to move forward: 
-read the datasheet and check if the code actually does the right thing
for the IPU restrictions.
-or just filter out the video modes that are problematic with the IPU 
unit.

Change-Id: Iab37b6119e3332e5c7a13570dcf855dec22ddb15