mxc_hdmi: Fix hang during boot
If HDMI has been in use by the bootloader or a previous
running kernel it hangs when HDMI cable is connected.

Move the enabling of overflow_into mxc_hdmi_setup() to fix this
issue.

Original patch is from Jon Nettleton based on a 3.10 kernel:
https://github.com/xbianonpi/xbian-package-kernel/blob/master/patches/imx6/hdmi-freeze.patch

Original patch is from Fabio Estevam:
https://github.com/Freescale/linux-fslc/commit/09845b886c1198c9461de424ffd95dde363324c1?diff=split

Change-Id: I532f03217274139db1740e68d1251d2ee5cf2b01